refactor: consolidate custom error handling for upscaling and video processing
- Removed redundant custom exception classes from video_player.py and ai_upscaling.py, replacing them with centralized imports from the exceptions module.
- Updated error handling in the AIUpscaler class to utilize the new UpscaleError for improved consistency and clarity in error reporting.
- Enhanced overall code organization by streamlining imports and reducing complexity in exception management, promoting better maintainability.
